+++ frontend/index.html
@@ -32,6 +32,10 @@
         </section>
       </section>
     </div>
-    <div id="desk-container"></div>
+    <div id="desk-container">
+      <!-- <img src="./assets/" alt="" srcset="" id="score-table"> -->
+      <!-- </img> -->
+      <img src="http://apper.co.il/wp-content/uploads/yanivThumb.jpg" id="Yaniv" >
+    </div>
   </body>
 </html>
