+++ frontend/style.scss
@@ -53,6 +53,18 @@ h1{
     width: 100%;
     height: 100%;
     
+    #score-table {
+        grid-row: 2;
+        grid-column: 2;
+    }
+
+    #Yaniv {
+        position: static;
+        grid-row: 4;
+        grid-column: 5;
+        border-radius: 50%;
+        border: 0.6rem solid black;
+    }
     .pile-deck{
         background-color: rgb(22, 19, 19);
         margin-right: 1rem;
